fit file import: split multisport files into multiple activities
Recently multisport files as produces by Garmin Fenix 3, some
Forerunners and so on are imported as a _Run_ activity that in turn is
wrong.

This commit adds functionality that was already merged with MR
https://github.com/GoldenCheetah/GoldenCheetah/pull/1397

The change splits multisport files into its embedded activities. This is
useful especially for combi trainings such as Duathlon, Triathlon, ...

The information lost is the details of the device type which is parsed
before any of the activities appear.
--> Multisport-tag - device types - {swim|transission|bike|run} tag -
data points.